Network.AWS.XRay.GetTraceSummaries

Description

Retrieves IDs and metadata for traces available for a specified time frame using an optional filter. To get the full traces, pass the trace IDs to BatchGetTraces .

A filter expression can target traced requests that hit specific service nodes or edges, have errors, or come from a known user. For example, the following filter expression targets traces that pass through api.example.com :

service("api.example.com")

This filter expression finds traces that have an annotation named account with the value 12345 :

annotation.account = "12345"

For a full list of indexed fields and keywords that you can use in filter expressions, see Using Filter Expressions in the AWS X-Ray Developer Guide .

Creating a Request

getTraceSummaries

Arguments

:: UTCTime
gtsStartTime

-> UTCTime
gtsEndTime

-> GetTraceSummaries

Creates a value of GetTraceSummaries with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsFilterExpression - Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

gtsNextToken - Specify the pagination token returned by a previous request to retrieve the next page of results.

gtsSampling - Set to true to get summaries for only a subset of available traces.

gtsStartTime - The start of the time frame for which to retrieve traces.

gtsEndTime - The end of the time frame for which to retrieve traces.

data GetTraceSummaries

See: getTraceSummaries smart constructor.

Request Lenses

gtsFilterExpression :: Lens' GetTraceSummaries (Maybe Text)

Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

gtsNextToken :: Lens' GetTraceSummaries (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of results.

gtsSampling :: Lens' GetTraceSummaries (Maybe Bool)

Set to true to get summaries for only a subset of available traces.

gtsStartTime :: Lens' GetTraceSummaries UTCTime

The start of the time frame for which to retrieve traces.

gtsEndTime :: Lens' GetTraceSummaries UTCTime

The end of the time frame for which to retrieve traces.

Destructuring the Response

getTraceSummariesResponse

Arguments

:: Int
gtsrsResponseStatus

-> GetTraceSummariesResponse

Creates a value of GetTraceSummariesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsrsTracesProcessedCount - The number of traces that were processed to get this set of summaries.

gtsrsNextToken - If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most most recent results, closest to the end of the time frame.

gtsrsApproximateTime - The start time of this page of results.

gtsrsTraceSummaries - Trace IDs and metadata for traces that were found in the specified time frame.

gtsrsResponseStatus - -- | The response status code.

data GetTraceSummariesResponse

See: getTraceSummariesResponse smart constructor.

Response Lenses

gtsrsTracesProcessedCount :: Lens' GetTraceSummariesResponse (Maybe Integer)

The number of traces that were processed to get this set of summaries.

gtsrsNextToken :: Lens' GetTraceSummariesResponse (Maybe Text)

If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most most recent results, closest to the end of the time frame.

gtsrsApproximateTime :: Lens' GetTraceSummariesResponse (Maybe UTCTime)

The start time of this page of results.

gtsrsTraceSummaries :: Lens' GetTraceSummariesResponse [TraceSummary]

Trace IDs and metadata for traces that were found in the specified time frame.

gtsrsResponseStatus :: Lens' GetTraceSummariesResponse Int

- | The response status code.