Network.AWS.XRay.Types

Description

Service Configuration

xRay :: Service

API version 2016-04-12 of the Amazon X-Ray SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is missing required parameters or has invalid parameters.

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeds the maximum number of requests per second.

Alias

data Alias

An alias for an edge.

See: alias smart constructor.

alias :: Alias

Creates a value of Alias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aNames - A list of names for the alias, including the canonical name.

aName - The canonical name of the alias.

aType - The type of the alias.

aNames :: Lens' Alias [Text]

A list of names for the alias, including the canonical name.

aName :: Lens' Alias (Maybe Text)

The canonical name of the alias.

aType :: Lens' Alias (Maybe Text)

The type of the alias.

AnnotationValue

data AnnotationValue

Value of a segment annotation. Has one of three value types: Number, Boolean or String.

See: annotationValue smart constructor.

annotationValue :: AnnotationValue

Creates a value of AnnotationValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avNumberValue - Value for a Number annotation.

avStringValue - Value for a String annotation.

avBooleanValue - Value for a Boolean annotation.

avNumberValue :: Lens' AnnotationValue (Maybe Double)

Value for a Number annotation.

avStringValue :: Lens' AnnotationValue (Maybe Text)

Value for a String annotation.

avBooleanValue :: Lens' AnnotationValue (Maybe Bool)

Value for a Boolean annotation.

BackendConnectionErrors

data BackendConnectionErrors

See: backendConnectionErrors smart constructor.

backendConnectionErrors :: BackendConnectionErrors

Creates a value of BackendConnectionErrors with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bceOtherCount -

bceTimeoutCount -

bceHTTPCode5XXCount -

bceConnectionRefusedCount -

bceHTTPCode4XXCount -

bceUnknownHostCount -

bceOtherCount :: Lens' BackendConnectionErrors (Maybe Int)

bceTimeoutCount :: Lens' BackendConnectionErrors (Maybe Int)

bceHTTPCode5XXCount :: Lens' BackendConnectionErrors (Maybe Int)

bceConnectionRefusedCount :: Lens' BackendConnectionErrors (Maybe Int)

bceHTTPCode4XXCount :: Lens' BackendConnectionErrors (Maybe Int)

bceUnknownHostCount :: Lens' BackendConnectionErrors (Maybe Int)

Edge

data Edge

Information about a connection between two services.

See: edge smart constructor.

edge :: Edge

Creates a value of Edge with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStartTime - The start time of the first segment on the edge.

eAliases - Aliases for the edge.

eResponseTimeHistogram - A histogram that maps the spread of client response times on an edge.

eReferenceId - Identifier of the edge. Unique within a service map.

eEndTime - The end time of the last segment on the edge.

eSummaryStatistics - Response statistics for segments on the edge.

eStartTime :: Lens' Edge (Maybe UTCTime)

The start time of the first segment on the edge.

eAliases :: Lens' Edge [Alias]

Aliases for the edge.

eResponseTimeHistogram :: Lens' Edge [HistogramEntry]

A histogram that maps the spread of client response times on an edge.

eReferenceId :: Lens' Edge (Maybe Int)

Identifier of the edge. Unique within a service map.

eEndTime :: Lens' Edge (Maybe UTCTime)

The end time of the last segment on the edge.

eSummaryStatistics :: Lens' Edge (Maybe EdgeStatistics)

Response statistics for segments on the edge.

EdgeStatistics

data EdgeStatistics

Response statistics for an edge.

See: edgeStatistics smart constructor.

edgeStatistics :: EdgeStatistics

Creates a value of EdgeStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esFaultStatistics - Information about requests that failed with a 5xx Server Error status code.

esOKCount - The number of requests that completed with a 2xx Success status code.

esTotalResponseTime - The aggregate response time of completed requests.

esErrorStatistics - Information about requests that failed with a 4xx Client Error status code.

esTotalCount - The total number of completed requests.

esFaultStatistics :: Lens' EdgeStatistics (Maybe FaultStatistics)

Information about requests that failed with a 5xx Server Error status code.

esOKCount :: Lens' EdgeStatistics (Maybe Integer)

The number of requests that completed with a 2xx Success status code.

esTotalResponseTime :: Lens' EdgeStatistics (Maybe Double)

The aggregate response time of completed requests.

esErrorStatistics :: Lens' EdgeStatistics (Maybe ErrorStatistics)

Information about requests that failed with a 4xx Client Error status code.

esTotalCount :: Lens' EdgeStatistics (Maybe Integer)

The total number of completed requests.

ErrorStatistics

data ErrorStatistics

Information about requests that failed with a 4xx Client Error status code.

See: errorStatistics smart constructor.

errorStatistics :: ErrorStatistics

Creates a value of ErrorStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eOtherCount - The number of requests that failed with untracked 4xx Client Error status codes.

eThrottleCount - The number of requests that failed with a 419 throttling status code.

eTotalCount - The total number of requests that failed with a 4xx Client Error status code.

eOtherCount :: Lens' ErrorStatistics (Maybe Integer)

The number of requests that failed with untracked 4xx Client Error status codes.

eThrottleCount :: Lens' ErrorStatistics (Maybe Integer)

The number of requests that failed with a 419 throttling status code.

eTotalCount :: Lens' ErrorStatistics (Maybe Integer)

The total number of requests that failed with a 4xx Client Error status code.

FaultStatistics

data FaultStatistics

Information about requests that failed with a 5xx Server Error status code.

See: faultStatistics smart constructor.

faultStatistics :: FaultStatistics

Creates a value of FaultStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fsOtherCount - The number of requests that failed with untracked 5xx Server Error status codes.

fsTotalCount - The total number of requests that failed with a 5xx Server Error status code.

fsOtherCount :: Lens' FaultStatistics (Maybe Integer)

The number of requests that failed with untracked 5xx Server Error status codes.

fsTotalCount :: Lens' FaultStatistics (Maybe Integer)

The total number of requests that failed with a 5xx Server Error status code.

HTTP

data HTTP

Information about an HTTP request.

See: hTTP smart constructor.

hTTP :: HTTP

Creates a value of HTTP with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

httpHTTPMethod - The request method.

httpHTTPStatus - The response status.

httpClientIP - The IP address of the requestor.

httpUserAgent - The request's user agent string.

httpHTTPURL - The request URL.

httpHTTPMethod :: Lens' HTTP (Maybe Text)

The request method.

httpHTTPStatus :: Lens' HTTP (Maybe Int)

The response status.

httpClientIP :: Lens' HTTP (Maybe Text)

The IP address of the requestor.

httpUserAgent :: Lens' HTTP (Maybe Text)

The request's user agent string.

httpHTTPURL :: Lens' HTTP (Maybe Text)

The request URL.

HistogramEntry

data HistogramEntry

An entry in a histogram for a statistic. A histogram maps the range of observed values on the X axis, and the prevalence of each value on the Y axis.

See: histogramEntry smart constructor.

histogramEntry :: HistogramEntry

Creates a value of HistogramEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

heCount - The prevalence of the entry.

heValue - The value of the entry.

heCount :: Lens' HistogramEntry (Maybe Int)

The prevalence of the entry.

heValue :: Lens' HistogramEntry (Maybe Double)

The value of the entry.

Segment

data Segment

A segment from a trace that has been ingested by the X-Ray service. The segment can be compiled from documents uploaded with PutTraceSegments , or an inferred segment for a downstream service, generated from a subsegment sent by the service that called it.

See: segment smart constructor.

segment :: Segment

Creates a value of Segment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sDocument - The segment document

sId - The segment's ID.

sDocument :: Lens' Segment (Maybe Text)

The segment document

sId :: Lens' Segment (Maybe Text)

The segment's ID.

ServiceId

data ServiceId

See: serviceId smart constructor.

serviceId :: ServiceId

Creates a value of ServiceId with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siAccountId -

siNames -

siName -

siType -

siAccountId :: Lens' ServiceId (Maybe Text)

siNames :: Lens' ServiceId [Text]

siName :: Lens' ServiceId (Maybe Text)

siType :: Lens' ServiceId (Maybe Text)

ServiceInfo

data ServiceInfo

Information about an application that processed requests, users that made requests, or downstream services, resources and applications that an application used.

See: serviceInfo smart constructor.

serviceInfo :: ServiceInfo

Creates a value of ServiceInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sState - The service's state.

sStartTime - The start time of the first segment that the service generated.

sRoot - Indicates that the service was the first service to process a request.

sResponseTimeHistogram - A histogram that maps the spread of service response times.

sDurationHistogram - A histogram that maps the spread of service durations.

sReferenceId - Identifier for the service. Unique within the service map.

sAccountId - Identifier of the AWS account in which the service runs.

sNames - A list of names for the service, including the canonical name.

sName - The canonical name of the service.

sEndTime - The end time of the last segment that the service generated.

sType - The type of service. * AWS Resource - The type of an AWS resource. For example, AWS::EC2::Instance for a application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used. * AWS Service - The type of an AWS service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table. * client - Represents the clients that sent requests to a root service. * remote - A downstream service of indeterminate type.

sEdges - Connections to downstream services.

sSummaryStatistics - Aggregated statistics for the service.

sState :: Lens' ServiceInfo (Maybe Text)

The service's state.

sStartTime :: Lens' ServiceInfo (Maybe UTCTime)

The start time of the first segment that the service generated.

sRoot :: Lens' ServiceInfo (Maybe Bool)

Indicates that the service was the first service to process a request.

sResponseTimeHistogram :: Lens' ServiceInfo [HistogramEntry]

A histogram that maps the spread of service response times.

sDurationHistogram :: Lens' ServiceInfo [HistogramEntry]

A histogram that maps the spread of service durations.

sReferenceId :: Lens' ServiceInfo (Maybe Int)

Identifier for the service. Unique within the service map.

sAccountId :: Lens' ServiceInfo (Maybe Text)

Identifier of the AWS account in which the service runs.

sNames :: Lens' ServiceInfo [Text]

A list of names for the service, including the canonical name.

sName :: Lens' ServiceInfo (Maybe Text)

The canonical name of the service.

sEndTime :: Lens' ServiceInfo (Maybe UTCTime)

The end time of the last segment that the service generated.

sType :: Lens' ServiceInfo (Maybe Text)

The type of service. * AWS Resource - The type of an AWS resource. For example, AWS::EC2::Instance for a application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used. * AWS Service - The type of an AWS service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table. * client - Represents the clients that sent requests to a root service. * remote - A downstream service of indeterminate type.

sEdges :: Lens' ServiceInfo [Edge]

Connections to downstream services.

sSummaryStatistics :: Lens' ServiceInfo (Maybe ServiceStatistics)

Aggregated statistics for the service.

ServiceStatistics

data ServiceStatistics

Response statistics for a service.

See: serviceStatistics smart constructor.

serviceStatistics :: ServiceStatistics

Creates a value of ServiceStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssFaultStatistics - Information about requests that failed with a 5xx Server Error status code.

ssOKCount - The number of requests that completed with a 2xx Success status code.

ssTotalResponseTime - The aggregate response time of completed requests.

ssErrorStatistics - Information about requests that failed with a 4xx Client Error status code.

ssTotalCount - The total number of completed requests.

ssFaultStatistics :: Lens' ServiceStatistics (Maybe FaultStatistics)

Information about requests that failed with a 5xx Server Error status code.

ssOKCount :: Lens' ServiceStatistics (Maybe Integer)

The number of requests that completed with a 2xx Success status code.

ssTotalResponseTime :: Lens' ServiceStatistics (Maybe Double)

The aggregate response time of completed requests.

ssErrorStatistics :: Lens' ServiceStatistics (Maybe ErrorStatistics)

Information about requests that failed with a 4xx Client Error status code.

ssTotalCount :: Lens' ServiceStatistics (Maybe Integer)

The total number of completed requests.

TelemetryRecord

data TelemetryRecord

See: telemetryRecord smart constructor.

telemetryRecord :: TelemetryRecord

Creates a value of TelemetryRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trSegmentsReceivedCount -

trSegmentsSentCount -

trSegmentsSpilloverCount -

trSegmentsRejectedCount -

trTimestamp -

trBackendConnectionErrors -

trSegmentsReceivedCount :: Lens' TelemetryRecord (Maybe Int)

trSegmentsSentCount :: Lens' TelemetryRecord (Maybe Int)

trSegmentsSpilloverCount :: Lens' TelemetryRecord (Maybe Int)

trSegmentsRejectedCount :: Lens' TelemetryRecord (Maybe Int)

trTimestamp :: Lens' TelemetryRecord (Maybe UTCTime)

trBackendConnectionErrors :: Lens' TelemetryRecord (Maybe BackendConnectionErrors)

Trace

data Trace

A collection of segment documents with matching trace IDs.

See: trace smart constructor.

trace :: Trace

Creates a value of Trace with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tId - The unique identifier for the request that generated the trace's segments and subsegments.

tSegments - Segment documents for the segments and subsegments that comprise the trace.

tDuration - The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

tId :: Lens' Trace (Maybe Text)

The unique identifier for the request that generated the trace's segments and subsegments.

tSegments :: Lens' Trace [Segment]

Segment documents for the segments and subsegments that comprise the trace.

tDuration :: Lens' Trace (Maybe Double)

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

TraceSummary

data TraceSummary

Metadata generated from the segment documents in a trace.

See: traceSummary smart constructor.

traceSummary :: TraceSummary

Creates a value of TraceSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tsAnnotations - Annotations from the trace's segment documents.

tsHasThrottle - One or more of the segment documents has a 429 throttling error.

tsUsers - Users from the trace's segment documents.

tsHasFault - One or more of the segment documents has a 500 series error.

tsServiceIds - Service IDs from the trace's segment documents.

tsIsPartial - One or more of the segment documents is in progress.

tsHasError - One or more of the segment documents has a 400 series error.

tsId - The unique identifier for the request that generated the trace's segments and subsegments.

tsHTTP - Information about the HTTP request served by the trace.

tsDuration - The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

tsResponseTime - The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

tsAnnotations :: Lens' TraceSummary (HashMap Text [ValueWithServiceIds])

Annotations from the trace's segment documents.

tsHasThrottle :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents has a 429 throttling error.

tsUsers :: Lens' TraceSummary [TraceUser]

Users from the trace's segment documents.

tsHasFault :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents has a 500 series error.

tsServiceIds :: Lens' TraceSummary [ServiceId]

Service IDs from the trace's segment documents.

tsIsPartial :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents is in progress.

tsHasError :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents has a 400 series error.

tsId :: Lens' TraceSummary (Maybe Text)

The unique identifier for the request that generated the trace's segments and subsegments.

tsHTTP :: Lens' TraceSummary (Maybe HTTP)

Information about the HTTP request served by the trace.

tsDuration :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

tsResponseTime :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

TraceUser

data TraceUser

Information about a user recorded in segment documents.

See: traceUser smart constructor.

traceUser :: TraceUser

Creates a value of TraceUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tuServiceIds - Services that the user's request hit.

tuUserName - The user's name.

tuServiceIds :: Lens' TraceUser [ServiceId]

Services that the user's request hit.

tuUserName :: Lens' TraceUser (Maybe Text)

The user's name.

UnprocessedTraceSegment

data UnprocessedTraceSegment

Information about a segment that failed processing.

See: unprocessedTraceSegment smart constructor.

unprocessedTraceSegment :: UnprocessedTraceSegment

Creates a value of UnprocessedTraceSegment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utsErrorCode - The error that caused processing to fail.

utsId - The segment's ID.

utsMessage - The error message.

utsErrorCode :: Lens' UnprocessedTraceSegment (Maybe Text)

The error that caused processing to fail.

utsId :: Lens' UnprocessedTraceSegment (Maybe Text)

The segment's ID.

utsMessage :: Lens' UnprocessedTraceSegment (Maybe Text)

The error message.

ValueWithServiceIds

data ValueWithServiceIds

Information about a segment annotation.

See: valueWithServiceIds smart constructor.

valueWithServiceIds :: ValueWithServiceIds

Creates a value of ValueWithServiceIds with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vwsiServiceIds - Services to which the annotation applies.

vwsiAnnotationValue - Values of the annotation.

vwsiServiceIds :: Lens' ValueWithServiceIds [ServiceId]

Services to which the annotation applies.

vwsiAnnotationValue :: Lens' ValueWithServiceIds (Maybe AnnotationValue)

Values of the annotation.